@@ -4,7 +4,7 @@
   "description": "",
   "main": "output_csv.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest",
     "build": "webpack --mode production --config webpack.config.js",
     "build-d": "webpack --mode development --config webpack.config.js"
   },
@@ -12,18 +12,26 @@
   "author": "",
   "license": "UNLICENSED",
   "devDependencies": {
+    "@babel/core": "^7.15.5",
+    "@babel/preset-env": "^7.15.6",
+    "babel-loader": "^8.2.2",
+    "eslint": "^7.32.0",
     "file-loader": "^6.0.0",
+    "jest": "^27.2.1",
+    "jest-esm-transformer": "^1.0.0",
     "url-loader": "^4.1.0",
-    "webpack": "^4.42.1",
+    "webpack": "^4.46.0",
     "webpack-cli": "^3.3.11"
   },
   "dependencies": {
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.5",
     "@holiday-jp/holiday_jp": "^2.3.0",
     "date-fns": "^2.13.0",
     "dayjs": "^1.8.28",
     "decimal.js": "^10.2.1",
     "encoding-japanese": "^1.0.30",
     "extensible-custom-error": "0.0.7",
+    "google-libphonenumber": "^3.2.23",
     "pdfmake": "^0.1.65"
   }
 }
