@@ -5,14 +5,32 @@ module.exports = (env, argv) => {
     if (IS_DEVELOPMENT) {
         return {
             mode: "development",
-            devtool:"source-map",
+            devtool: "source-map",
             module: {
                 rules: [
                     {
                         test: /\.png$/i,
                         use: ["url-loader"]
-                    }
-                ]
+                    },
+                    {
+                        test: /\.(js|jsx)$/,
+                        exclude: /node_modules/,
+                        use: {
+                            loader: "babel-loader",
+                            options: {
+                                presets: [
+                                    ["@babel/preset-env", { targets: "defaults" }]
+                                ],
+                                plugins: [
+                                    [
+                                        "@babel/plugin-proposal-nullish-coalescing-operator",
+                                        { loose: false }
+                                    ],
+                                ],
+                            },
+                        },
+                    },
+                ],
             },
 
             entry: {
@@ -25,7 +43,7 @@ module.exports = (env, argv) => {
                 insert_collect_credit_button: "./insert_collect_credit_button.js",
                 calc_member_fee: "./app_apply/button_calc_member_fee.js",
                 button_wfi_antisocial_check: "./app_apply/button_wfi_antisocial_check.js",
-                button_wfi_add_kyoryoku_master: "./app_apply/button_wfi_add_kyoryoku_master.js",
+                button_add_kyoryoku_master: "./app_apply/button_add_kyoryoku_master.js",
                 button_make_all_payment_detail_checked: "./app_apply/button_make_all_payment_detail_checked.js",
                 button_make_records_ready_to_pay: "./app_apply/button_make_records_ready_to_pay.js",
                 app_collect: "./app_collect.js",
@@ -54,8 +72,26 @@ module.exports = (env, argv) => {
                     {
                         test: /\.png$/i,
                         use: ["url-loader"]
-                    }
-                ]
+                    },
+                    {
+                        test: /\.(js|jsx)$/,
+                        exclude: /node_modules/,
+                        use: {
+                            loader: "babel-loader",
+                            options: {
+                                presets: [
+                                    ["@babel/preset-env", { targets: "defaults" }]
+                                ],
+                                plugins: [
+                                    [
+                                        "@babel/plugin-proposal-nullish-coalescing-operator",
+                                        { loose: false }
+                                    ],
+                                ],
+                            },
+                        },
+                    },
+                ],
             },
 
             entry: {
@@ -68,7 +104,7 @@ module.exports = (env, argv) => {
                 insert_collect_credit_button: "./insert_collect_credit_button.js",
                 calc_member_fee: "./app_apply/button_calc_member_fee.js",
                 button_wfi_antisocial_check: "./app_apply/button_wfi_antisocial_check.js",
-                button_wfi_add_kyoryoku_master: "./app_apply/button_wfi_add_kyoryoku_master.js",
+                button_add_kyoryoku_master: "./app_apply/button_add_kyoryoku_master.js",
                 button_make_all_payment_detail_checked: "./app_apply/button_make_all_payment_detail_checked.js",
                 button_make_records_ready_to_pay: "./app_apply/button_make_records_ready_to_pay.js",
                 app_collect: "./app_collect.js",
